change data structure

+++ c_to_f.py
@@ -1,4 +1,4 @@
 c = input('請輸入攝氏溫度: ')
-c = int(c)
+c = float(c)
 f = c * 5 / 9 + 32
 print(f)
